N Queens
This program is an example of the N Queens Problem
in which given a normal Chess board, the program has to
figure the most optimal way to place N Queens such that no
Queens check one another.

--Meant to be implemented using an Evolutionary Algorithm.  Current
version uses Mutation to alter Genotype.  However, every position in the
genotype is mutated so it isn't necessarily using optimal evolutionary
strategy.